refactor: local preview mode refactor and fix missing config error (#156)
