+++ website/app/utils/preview/hash-change.ts
@@ -1,0 +1,21 @@
+import { useEffect, useState } from 'react';
+
+export function useHashChange(): string {
+  const [hash, setHash] = useState('');
+
+  function onHashChange() {
+    // TODO: probably have to handle deeper index files, and redirects somewhere
+    const newHash =
+      window.location.hash.replace('#', '') === '/'
+        ? '/index'
+        : window.location.hash.replace('#', '');
+    return setHash(newHash);
+  }
+
+  useEffect(() => {
+    window.addEventListener('hashchange', onHashChange);
+    return () => window.removeEventListener('hashchange', onHashChange);
+  }, []);
+
+  return hash;
+}
